[transform] move test classes into test target
